+++ TinyNvidiaUpdateChecker/MainConsole.cs
@@ -76,11 +76,6 @@ namespace TinyNvidiaUpdateChecker
         private static int osID;
 
         /// <summary>
-        /// DCH ID, 1 indicates that the system requires DCH drivers, and 0 standard drivers
-        /// </summary>
-        private static int dchID = 0;
-
-        /// <summary>
         /// Show UI or go quiet mode
         /// </summary>
         public static bool showUI = true;
@@ -562,20 +557,9 @@ namespace TinyNvidiaUpdateChecker
                 pfID = 756; // GTX 970
             }
 
-            // Check if system requires DCH drivers, thanks to https://github.com/Osspial
-            try {
-                using (var regKey = Registry.LocalMachine.OpenSubKey(@"SYSTEM\CurrentControlSet\Services\nvlddmkm", false)) {
-                    if (regKey != null) {
-                        if (regKey.GetValue("DCHUVen") != null) {
-                            dchID = 1;
-                        }
-                    }
-                }
-            } catch { }
-
             // finish request
             try {
-                gpuURL = $"https://www.nvidia.com/Download/processDriver.aspx?psid={psID}&pfid={pfID}&osid={osID}&lid={langID}&dtcid={dchID}&ctk=0";
+                gpuURL = $"https://www.nvidia.com/Download/processDriver.aspx?psid={psID}&pfid={pfID}&osid={osID}&lid={langID}&dtcid=1&ctk=0";
 
                 WebClient client = new WebClient();
                 Stream stream = client.OpenRead(gpuURL);
@@ -700,6 +684,7 @@ namespace TinyNvidiaUpdateChecker
             }
 
             if (debug) {
+                Console.WriteLine($"gpuURL: {gpuURL}");
                 Console.WriteLine($"downloadURL: {downloadURL}");
                 Console.WriteLine($"pdfURL:      {pdfURL}");
                 Console.WriteLine($"releaseDate: {releaseDate.ToShortDateString()}");