+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Non-DCH drivers are no longer supported by NVIDIA. So they are always opted to be installed. See issue #101
+
 ## [1.14.5] - 2021-09-06
 ### Added
 - DCH driver support